Config Updates
- Changed activity configuration;
- MainUserScreen is now MainActivity;
- Signup screen is now StartActivity;
