+++ app/src/main/java/ie/ul/tutorfinder/LoginActivity.java
@@ -49,7 +49,7 @@ public class LoginActivity extends AppCompatActivity {
                             @Override
                             public void onComplete(@NonNull Task<AuthResult> task) {
                                 if (task.isSuccessful()) {
-                                    Intent mainIntent = new Intent( LoginActivity.this, MainUserScreen.class );
+                                    Intent mainIntent = new Intent( LoginActivity.this, MainActivity.class );
                                     mainIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
                                     startActivity(mainIntent);
                                     finish();
