+++ app/src/main/res/layout/activity_login.xml
@@ -7,11 +7,20 @@
     android:orientation="vertical"
     tools:context="LoginActivity">
 
+    <TextView
+        android:id="@+id/textView6"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginLeft="30dp"
+        android:layout_marginTop="50dp"
+        android:text="Please Login!"
+        android:textSize="24sp" />
 
     <EditText
         android:id="@+id/etUserEmail"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginTop="30dp"
         android:ems="10"
         android:hint="Email"
         android:inputType="textEmailAddress" />
@@ -29,5 +38,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="Login" />
+
 </LinearLayout>
